\name{icsw-package}
\alias{icsw-package}
\alias{icsw}
\docType{package}
\title{
Inverse compliance score weighting
}
\description{
Inverse compliance score weighting is a tool for estimating 
average treatment effects from local average treatment 
effects by weighting units using inverse probabilities of 
complying with an encouragement (instrument).
}
\details{
\tabular{ll}{
Package: \tab icsw\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2018-05-16\cr
License: \tab  MIT\cr
LazyLoad: \tab yes\cr
}
}

\author{
Peter M. Aronow <peter.aronow@yale.edu>; Dean Eckles <icsw@deaneckles.com>; Kyle Peyton <kyle.peyton@yale.edu>

Maintainer: Kyle Peyton <kyle.peyton@yale.edu>
}
\references{

Bethany Albertson and Adria Lawrence. (2009). After the credits roll: The long-term effects of educational television on public knowledge and attitudes. \emph{American Politics Research}. 37(2): 275-300.

Peter M. Aronow and Allison Carnegie. (2013). Beyond LATE: Estimation of the average treatment effect with an instrumental variable. \emph{Political Analysis}. 21.4 (2013): 492-506.

Peter M. Aronow and Allison Carnegie. (2013). Replication data for: Beyond LATE: Estimation of the average treatment effect with an instrumental variable. \emph{Dataverse Network.} http://hdl.handle.net/1902.1/21729 (accessed May 14, 2015).
}
\keyword{ package }

